pyproject.toml
[project]
name = "viralforge"
version = "0.1.0"
description = "Trend-aware original animated YouTube video automation with copyright and policy guardrails."
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "feedparser>=6.0.11",
  "google-api-python-client>=2.160.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.1",
  "moviepy>=1.0.3,<2.0",
  "numpy>=1.26.4",
  "Pillow>=10.4.0",
  "edge-tts>=7.0.0",
  "pyttsx3>=2.90",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "rich>=13.9.0",
  "tenacity>=9.0.0"
]

[project.scripts]
viralforge = "viralforge.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
